# Yarp/wwwroot/admin/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "One Point Admin API",
    "description": "Admin APIs to configure settings.",
    "termsOfService": "https://www.onepointfinserv.com/",
    "contact": {
      "name": "OnePoint Financial Services",
      "url": "https://www.onepointfinserv.com/",
      "email": "anuj@onepointfinserv.com"
    },
    "license": {
      "name": "OnePoint",
      "url": "https://www.onepointfinserv.com/"
    },
    "version": "1.0.0"
  },
  "paths": {
...
    "/config/uploadFile": {
      "post": {
        "tags": [
          "Admin"
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              },
              "encoding": {
                "file": {
                  "style": "form"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/config/adminApi": {
      "post": {
        "tags": [
          "Admin"
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              },
              "encoding": {
                "file": {
                  "style": "form"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/config/clusters": {
      "post": {
        "tags": [
          "Cluster"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateClusterConfigModel"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateClusterConfigModel"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateClusterConfigModel"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "get": {
        "tags": [
          "Cluster"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/config/clusters/{clusterId}": {
      "get": {
        "tags": [
          "Cluster"
        ],
        "parameters": [
          {
            "name": "clusterId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "put": {
        "tags": [
          "Cluster"
        ],
        "parameters": [
          {
            "name": "clusterId",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "clusterId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PutClusterConfigModel"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/PutClusterConfigModel"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/PutClusterConfigModel"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "patch": {
        "tags": [
          "Cluster"
        ],
        "parameters": [
          {
            "name": "clusterId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PutClusterConfigModel"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/PutClusterConfigModel"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/PutClusterConfigModel"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/setup/consumer": {
      "post": {
        "tags": [
          "Consumer"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateConsumerModel"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateConsumerModel"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateConsumerModel"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "get": {
        "tags": [
          "Consumer"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/setup/consumer/{consumerId}": {
      "put": {
        "tags": [
          "Consumer"
        ],
        "parameters": [
          {
            "name": "consumerId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateConsumerModel"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateConsumerModel"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateConsumerModel"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "patch": {
        "tags": [
          "Consumer"
        ],
        "parameters": [
          {
            "name": "consumerId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatchConsumerModel"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/PatchConsumerModel"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/PatchConsumerModel"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "get": {
        "tags": [
          "Consumer"
        ],
        "parameters": [
          {
            "name": "consumerId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "Consumer"
        ],
        "parameters": [
          {
            "name": "consumerId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/setup/consumerGroup": {
      "post": {
        "tags": [
          "ConsumerGroup"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateConsumerGroupModel"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateConsumerGroupModel"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateConsumerGroupModel"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "get": {
        "tags": [
          "ConsumerGroup"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/setup/consumerGroup/{consumerGroupId}": {
      "put": {
        "tags": [
          "ConsumerGroup"
        ],
        "parameters": [
          {
            "name": "consumerGroupId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateConsumerGroupModel"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateConsumerGroupModel"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateConsumerGroupModel"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "patch": {
        "tags": [
          "ConsumerGroup"
        ],
        "parameters": [
          {
            "name": "consumerGroupId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateConsumerGroupModel"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateConsumerGroupModel"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateConsumerGroupModel"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "get": {
        "tags": [
          "ConsumerGroup"
        ],
        "parameters": [
          {
            "name": "consumerGroupId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "ConsumerGroup"
        ],
        "parameters": [
          {
            "name": "consumerGroupId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/credential/basic": {
      "post": {
        "tags": [
          "Credential"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BasicCredModel"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/BasicCredModel"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/BasicCredModel"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/credential/basic/{basicCredId}": {
      "put": {
        "tags": [
          "Credential"
        ],
        "parameters": [
          {
            "name": "basicCredId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BasicCredModel"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/BasicCredModel"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/BasicCredModel"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "BasicCredModel": {
        "type": "object",
        "properties": {
          "consumerId": {
            "type": "string",
            "nullable": true
          },
          "username": {
            "type": "string",
            "nullable": true
          },
          "password": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ClusterHttpClientConfigModel": {
        "type": "object",
        "properties": {
          "sslProtocols": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "dangerousAcceptAnyServerCertificate": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
      "CreateClusterConfigModel": {
        "type": "object",
        "properties": {
          "destinationAddress": {
            "type": "object",
            "additionalProperties": {
              "type": "string",
              "nullable": true
            },
            "nullable": true
          },
          "loadBalancingPolicy": {
            "type": "string",
            "nullable": true
          },
          "httpClient": {
            "$ref": "#/components/schemas/ClusterHttpClientConfigModel"
          }
        },
        "additionalProperties": false
      },
      "CreateConsumerGroupModel": {
        "required": [
          "name"
        ],
        "type": "object",
        "properties": {
          "name": {
            "minLength": 1,
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "CreateConsumerModel": {
        "required": [
          "consumerGroupId",
          "name"
        ],
        "type": "object",
        "properties": {
          "consumerGroupId": {
            "minLength": 1,
            "type": "string"
          },
          "name": {
            "minLength": 1,
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "PatchConsumerModel": {
        "type": "object",
        "properties": {
          "consumerGroupId": {
            "type": "string",
            "nullable": true
          },
          "name": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "PutClusterConfigModel": {
        "type": "object",
        "properties": {
          "destinationAddress": {
            "type": "object",
            "additionalProperties": {
              "type": "string",
              "nullable": true
            },
            "nullable": true
          },
          "loadBalancingPolicy": {
            "type": "string",
            "nullable": true
          },
          "httpClient": {
            "$ref": "#/components/schemas/ClusterHttpClientConfigModel"
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "apiKey",
        "description": "Key based authentication based on resource using various schemes.",
        "name": "Authorization",
        "in": "header"
      }
    }
  },
  "security": [
    {
      "Bearer": []
    }
  ]
}